refactor: move playwright tracing logic to its own class
